Fix error when an item is moved to a position of a disappearing item.
Summary:
When we move the item to the root host (for disappearing animations) we use the same index the item had for the previous tree. The problem with this is that when running unmountOrMoveOldItems it can happen that an other mount item is moved to that index before we call the unmount delegate. When we move an item in ComponentHost, we add that to a scrap list.
The fix is we'll try to get the mount item from the mMountItems list in ComponentHost and if we don't find it, it has to be in the mScrapMountItemsArray.

This is how the order of operations can create this issue:
1. Update state: new tree removes a disappearing component (index **4**) and puts a new one moving an already mounted component in the position of the disappearing one (move from **3** to **4**).
2. On before mount we collect disappearing items and move them to the root component if needed (also add them to a list to hook the unmount). When we add the item to the root we use the same index, in this case **4**.
3. When we run unmountOrMoveOldItems a previous mounted item in position **3** needs to move to position **4** because the new tree adds a component before it. For this we call ComponentHost.moveItem(item, 3, 4)
    - Internally this will call ComponentHostUtils.moveItem(oldIndex, newIndex, mMountItems, mScrapMountItemsArray) (and their respected view or drawables lists)
    - In this case, the component in position **4** will be added to mScrapMountItemsArray and removed from mMountItems as the mScrapMountItemsArray does not contain an item in that index.
    - The next loop, which will test index 4, is where we find the disappearing item. With disappearing items we don't use the index to start unmounting them as we can not be sure if they will match the one assigned in the new pass (in this example they match but it's not always the case, this is specially true for RenderCore), so we need to find the index in the list of mounted items. **The fix** here is that we should look for the index in either the mMountItems array or in mScrapMountItemsArray as it may have been moved (like we see in this example).
    - With the index we then call ComponentHostUtils.removeItem(index, mMountItems, mScrapMountItemsArray) where it would first check mScrapMountItemsArray and then mMountItems [link](https://www.internalfb.com/intern/diffusion/FBS/browse/master/fbandroid/libraries/components/litho-core/src/main/java/com/facebook/litho/ComponentHostUtils.java?commit=ae3e1bcfb581880a15d82b39667b206509cf94fe&lines=90%2C97%2C91-96).

This fixes: [link](https://www.internalfb.com/intern/logview/details/facebook_android_crashes/31395f3e75e761fb54358d909dd5d8be/?selected-logview-tab=latest&drillstate=%7B%22start%22%3A%22-1+week%22%2C%22end%22%3A%22now%22%2C%22context%22%3A%22facebook_android_crashes%22%2C%22constraints%22%3A%5B%7B%22col%22%3A%22app_version%22%2C%22op%22%3A%22in%22%2C%22vals%22%3A%5B%22280.0.0.26.122%22%5D%7D%2C%7B%22col%22%3A%22mid%22%2C%22op%22%3A%22).

Reviewed By: mihaelao

Differential Revision: D22666369

fbshipit-source-id: 982a3fed5ec1b1dbb73c74e04e28a0c80e34f163