Makes event classes without @Event annotation
Summary:
In order move the visibility event classes from Litho to RenderCore `Event` annotation has to become optional.

This diff makes the `EventDeclarationsExtractor` return void if the event class used does not have the `Event` annotation. As side effect this enables the use of event classes without the annotation as long as they do not need to return from the event.

Reviewed By: marco-cova

Differential Revision: D22791206

fbshipit-source-id: e406ace109257a6d14150a8fbb5aacc4ebf7a5cf