Add a componentTag to components to allow arbitrary data to be attached for debugging / testing. (#680)
Summary:
It is currently not possible in Litho to attach arbitrary information to Components for use in debugging and testing. This information can be incredibly useful for debugging, for instance if it is the data source of props on the component. The changes proposed in this PR would enable access to this underlying data source when debugging through Android Studio, and can potentially allow this information to be exposed to tools like Flipper or Stetho's UI inspector.

## Changelog

Add a componentTag to components to allow arbitrary data to be attached for debugging / testing.

Pull Request resolved: https://github.com/facebook/litho/pull/680

Test Plan: Added test to CommonPropsTest to verify the underlying InternalNode has its componentTag set. `./gradlew clean test` passes successfully. Additionally verified manually in a project which uses Litho.

Reviewed By: topwu

Differential Revision: D22833106

Pulled By: adityasharat

fbshipit-source-id: 83c43bb796cbaac94a8d732e6360b43ee27116c8