Add logic into TextInputSpec to more consistently show keyboard on requestFocus
Summary: I'm working on some samples to show how to show the keyboard on first mount. Unfortunately it's very hard to get right because of InputMethodManager + focus beng quite flakey in android. This attempts to fix this by posting a Runnable when necessary (which is similar to what Android's SearchView does).

Reviewed By: Katalune

Differential Revision: D22921269

fbshipit-source-id: f5a6ddf5cb11a534552149eafb3673cb9bbe8e79