IncrementalMountExtensions: Unmount off-screen items for dirty mounts
Summary:
For dirty mounts, if an already mounted item is moved off-screen, we can't rely on MountState to unmount it after the ref count is decreased, because MountState only mounts or updates items (by default it mounts everything, we only have the option to skip mounting certain items. It does not check if an already mounted item has become unmountable).
We need to trigger unmount on these items on a dirty mount if the ref count is 0, but we need to do it in `afterMount` - otherwise if in the new layout state certain items moved position, we can end up marking the wrong item for unmounting. We need to wait for prepareMount first, which moves the existing mounted items to the new positions if needed.

Reviewed By: adityasharat

Differential Revision: D22979825

fbshipit-source-id: 8aa938d1a2119d87d6d49f6b863dd90e7be8badd