Fixes a bugs in isReconcilable
Summary:
Fixes a bugs in isReconcilable. This change set fixes a bug where reconciliation would not be invoked when a delegating components (like Wrapper) is at the root and the delegate is a custom component (like Column) or a layout spec which have some props which are Components.

Now the utils class accounts of this case and the custom components will call the `isEquivalentTo(Component, boolean)` API.

Reviewed By: astreet

Differential Revision: D22999170

fbshipit-source-id: 9332e1200aeafd42264a62b2606ebf1b1078154d