Fix cached value generation to respect global key
Summary: A cached value should be scoped to the specific component (and global key). The current logic would cause two components of the same type in a single ComponentTree to reuse the same cached value between them. Since this is often used for Drawables, this would cause the same drawable to be mounted to multiple different views. The solution here is to add the globalKey as a parameter to the cached value lookup.

Reviewed By: mihaelao

Differential Revision: D23004220

fbshipit-source-id: f4aea76083b38fe93d0697b8af40c0a3ebb5cec1