Improve onError propagation
Summary:
Improve onError propagation for OnError callbacks

The `ReThrownException`  should be the only formal way to re-raise exceptions, that is abstracted into `ComponentUtils.raise(ComponentContext, Exception)`. Specs should use `ComponentUtils.raise(ComponentContext, Exception)` to re-throw exceptions from their `OnError` callback.

`Layout` has a modified error handling.

When encountering a `ReThrownException` it is guaranteed that the exception was raised from an `OnError` call, so all operations should be done on the `original` exception. This is abstracted into `Layout#handle(ComponentContext, Component, ReThrownException)` .

When encountering a `ComponentChainException` it may have been previously re-raised; after checking if a current error handler is different from the parent error handler, the exception is dispatched to the parent handler. Here the `original` exception is not used, since the spec might want to read component chain. This is abstracted into `Layout#handle(ComponentContext, Component, ComponentChainException)`.

If the error handler re-raises the exception, it is caught, wrapped in a ComponentChainException, and thrown again. This allows the component chain to updated as the error is propagated upwards.

If the and ErrorEventHandler is encountered, then exception has been propagated up to the root, so the exceptions is wrapped in a ComponentChainException, and thrown again to capture the root component in the chain.

In addition this also fixes a bug where nested trees were resolved using incorrect parent context.

Reviewed By: arpitratan

Differential Revision: D23241809

fbshipit-source-id: a622ebf5632a3e40c527dc4bb490238ee8370c15