Pass scoped context to dispatch working range enter/exit
Summary:
In the generated code for dispatchOnEnteredRange/dispatchOnExitedRange, we get the scoped ComponentContext from the Component.
We already have access to it when these methods are called, so we can pass it in.
This is needed as part of making Component stateless, which means removing the ComponentContext scope field from the Component class.

Reviewed By: adityasharat

Differential Revision: D23318658

fbshipit-source-id: 5ed3871a92a085ed5e6a6c1aba6acc7da3addcf2