Remove RenderState usage from BloksLithoComponent
Summary:
This fixes one of the last places in which we were nesting RenderState inside another tree.
This new approach means that we can commit the layout result coming from RenderCore together with the Litho layout. This reduces complexity and makes our commit strategy correct (before this we could only guarantee it to be "eventually correct")

Reviewed By: colriot

Differential Revision: D22999988

fbshipit-source-id: 9e4edcada38e06904464443c58e6afec88c2837a