Add support to read/write from field-less data
Summary:
Editors are designed to work on fields as they provide rich information about the environment where the data is, such as field name and holder class. It also allows updating field references for immutable values.

There are cases where data may not have a field reference where the data is held, such as with Collections. These helpers are to bridge that gap by using a transient class to read and write that value.

Reviewed By: Katalune

Differential Revision: D23564107

fbshipit-source-id: f645b4356f471088c9e67fa071409b85c12d54cb