Implements host translations
Summary:
Context:
There is a fundamental difference between Litho and RenderCore: **Lithos' bounds are absolute where as RC's bounds are relative**. `hostTranslationX`, and `hostTranslationX` were added to `RenderTreeNode` to enable Litho on RenderCore to correctly process visibility events, and do incremental mount.

Litho, currently, hijacks `RenderTreeNode` creation (reduce phase) from RC, so for now, Litho forces absolute bounds on the `RenderTreeNode`, and uses `hostTranslationX` and `hostTranslationX` in RenderCore to correctly mount items. **Litho's usage of `RenderTreeNode` is a hack**.

This change set achieves 2 things

* Isolates the hack within Litho as much as possible.

  * Makes the hack-y constructor of `RenderTreeNode` private and deprecated.

  * Use a new static deprecated utility method to create instances only for Litho on RenderCore.

* Adds the absolute x and y position for every `RenderTreeNode`.

  * (**THE REAL REASON FOR THIS DIFF**) Layout extensions (visitors, like visibility and incremental mount) need the absolute bounds.

  * Adds an API to formally get the absolute bounds from a `RenderTreeNode`.

  * We don't necessarily need absolute x and y set in each `RenderTreeNode`, it can be calculated at runtime, but that is likely to be an expensive operation (by traversing upwards to the root via the parent).

In order to remove the deprecated methods the usage of `RenderTreeNode#getBounds()` in Litho and replace it with `RenderTreeNode#getAbsoluteBounds()` (which is done in D23376438)

Reviewed By: mihaelao

Differential Revision: D23241761

fbshipit-source-id: e4795a2712968048a5e0bdd08efcd8b2f4d78e73