Add litho-editor-flipper
Summary:
At the moment Flipper for Litho depends on Litho editor in a class FlipperEditor that lives in the Flipper repo. FlipperEditor has a stable read/write API and is modified internally when we make changes to support new cases on the editor.

In this diff I'm moving this FlipperEditor class to the Litho repo into a new dependency called litho-editor-flipper (note that the editor is litho-editor-core). That allows us to update it locally without worrying about Flipper OSS, and the changes will propagate to Flipper on Litho's OSS schedule.

Reviewed By: colriot

Differential Revision: D23379387

fbshipit-source-id: ef88c7cc9f3388e5a2f7627a948352213bfb0361