Don't set scoped ComponentContext on background/foreground Drawables
Summary:
There's no reason we need a scoped context on the background/foreground Drawables -  it was originally added to be cautious about NPEs in D3662214.
We should never be using the scoped context for these components.

They are never assigned a global key either, so as we move to fetch the scoped context from the LayoutState based on key, it will be null.

Reviewed By: adityasharat

Differential Revision: D23574546

fbshipit-source-id: 1576c6c1a8d543511b5ac89c90faaa79c2cc528e