Remove getScopedContext in Triggers generated code
Summary:
In some of the methods generated for Triggers, we access the scoped context from the Component or Section instance. Since we're trying to remove that field, we can pass in the scoped ComponentContext instead of accessing it from the Component.
Right now we pass it when we bind the trigger with the Component  through getComponentScope. In the upcoming diffs, we'll replace this to fetch the ComponentContext from the key:ComponentContext LayoutState-scoped map.

Reviewed By: adityasharat

Differential Revision: D23398276

fbshipit-source-id: 14d80cd08dd984b1926bda26c79962bd816cdbd9