Pass scoped context to getCachedValue methods
Summary:
Right now, the generated `getCachedValue` implementation fetches the scoped ComponentContext from the Component instance. This diff adds the context as param to the generated `getCachedValue` method, as part of making `getScopedContext` on the Component obsolete.
The  methods using CachedValue fields have the scoped context passed to them, so they can pass it to `getCachedValue`.

Reviewed By: adityasharat

Differential Revision: D23318667

fbshipit-source-id: 5005a9b445655bf901e90619cb333c489bda9193