Replace getScopedContext usages with passing ComponentContext
Summary:
There are a few usages of `component.getScopedContext` which can be replaced with passing in the scoped ComponentContext. This replaces those usages by passing in the ComponentContext directly.
Where necessary, `component.getScopedContext` calls have been moved to where we have access to the global key - in future diffs we'll use this global key to fetch the context from the LayoutState rather than from the Component.

Reviewed By: adityasharat

Differential Revision: D23318668

fbshipit-source-id: c3b57ee59ae9ea0a5c731fa68b4fc45a84392efa