Account for overflow in tops/bottoms LayoutState comparators
Summary:
The tops/bottoms comparators in LayoutState are causing crashes on certain inputs with the error "Comparison method violates its general contract!". There are tests to assert that the comparator is reflexive, symmetric and transitive, the issue is that the implementation can cause an overflow when tops/bottoms are extreme values and return incorrect results from the comparison.

Example:
Comparing a LayoutOutput with top Integer.MAX_VALUE with a LayoutOutput with top Integer.MIN_VALUE would return lhsTop - rhsTop which overflows and is a negative number.

This alternative implementation eliminates the subtraction to avoid overflow errors.

Reviewed By: adityasharat

Differential Revision: D23814405

fbshipit-source-id: 3b7d71f7c46fbea958bd19fa0af9e63113ddea4b