Adds RenderCoreExtensionHost
Summary:
*The current name of the interface is a placeholder; open to other suggestions.*

All the three extensions (visibility, transition, and incremental mount) "need" an API to inform inner/nested `RootHost` and `RenderTreeHost` that their bounds have changed.

The origin of this API (or it use) from Litho; where if a LithoView is hosting LithoViews, and the framework incrementally mount the parent LithoView, then we need to notify the descendants that their bounds have changed to process visibility events, fire transitions, and perform incremental mount.

This change-set adds an interface which `RootHost` and `RenderTreeHost` are now extending, allowing RC  (or an extension) to check if a content hosts a `RenderTree` and notify it of any potential changes in its visible bounds.

Reviewed By: mihaelao

Differential Revision: D23815903

fbshipit-source-id: 347515c0415715588fc5105f3685597bc6ecf33c