Refactor out the use of LithoView in IncrementalMountExtension
Summary:
This change-set removes the usage of LithoView from the IncrementalMountExtension.

`LithoView` was only required to have access to an API to notify child LithoViews about potentially changing visible bounds. This has been abstracted into the new RC API `RenderCoreExtensionHost` (name is tentative).

In addition, the IncrementalMountExtension was using a Litho Component API (`isMountViewSpec`) to exit early. This change-set simply uses an `instanceOf` check as a proxy for that.

The `mountingAll` feature implemented by passing in a custom bounds equal to the size of the view has been removed as it seems to a special case handling for Litho. It is left to the descendant host to handle changes in its visible bounds.

Reviewed By: mihaelao

Differential Revision: D23815901

fbshipit-source-id: 591869557d8717e9a567f1669649205ebb080ff9