Reset pressed state for HostView on unmount
Summary:
Sometime a view is not getting it's 'pressed' state reset before unmount, causing that state to not be cleared and carried to next reuse, therefore applying wrong drawable state.
Particular case where this might happen is when view is unmounted as soon as click event is triggered.
This is very similar to an issue we had in Litho, see D13675335 (https://github.com/facebook/litho/commit/1c8ad4b9c3ce9480efd631067bc9ccc2b431bc82) for more details.

Reviewed By: pasqualeanatriello

Differential Revision: D23951575

fbshipit-source-id: 65a022dffdbc0e286afb07f5c328d88c54da27c4