Move Transition related code to its own project [part 1]
Summary:
Here we are moving the code from Litho to the transition project.
We are keeping the same packages for old clases as it would otherwise require a big refactor (also oss would break).

This is missing the several classes (that will be moved in part 2) as they would require this already extracted to work.

Reviewed By: mihaelao

Differential Revision: D23575050

fbshipit-source-id: f7b576cf28c88b1c0b8f3fdd527bcd7a16661a01