Fix problem with calling moveItem in cases it was not necessary (and with same from/to index)
Summary:
We would be moving things inside the Host if their position in the flat array changed but that didn't necessarily mean that their position in the parent host had changed (this is actually always true for Litho instead).
In those cases we were trying to move things from index idx to index idx, making the host go in some undefined state.
This diff fixes the behaviour and adds a test.

Reviewed By: adityasharat

Differential Revision: D23987764

fbshipit-source-id: 3a553ec5096ed05cc80a4a6d29cb729dee0652bc