Pass key to bind event handlers
Summary:
LayoutState creates a list of Components during each layout pass to bind to event handlers. During `bindEventAndTriggerHandlers`, we iterate over this list and re-bind the existing event handlers with the updated Components and their scoped contexts.
In order to access the scoped context, we'll need the global key which we don't want to get from the Component instance anymore - so we save a map of global keys to Components instead.

Eventually we won't need the list of Components at all, we can get everything from the list of keys and the LayoutStateContext.

Reviewed By: astreet

Differential Revision: D23348700

fbshipit-source-id: d0c1ce75ac41fae52228fc33d0fe42c5a94b955a