Pass global key where a scoped context is used in DefaultInternalNode
Summary:
From this point on in the stack, we're working towards fetching the scoped ComponentContext from the LayoutStateContext based on the global key, instead of reading the Component scope field.
We don't want to read the global key from the Component either, since we want to remove that at well.
So we need to make sure that in places where we need to get the scoped context we have the global key available. This passes the global key to the DefaultInternalNode methods which need a scoped context

Reviewed By: adityasharat

Differential Revision: D23348675

fbshipit-source-id: 48d254053c4f484a2dfc2b1ef5980c61149cc131