Add getComponentKeysNeedingPreviousRenderData()
Summary:
LayoutState exposes `getComponentKeysNeedingPreviousRenderData` as a list of Components which it accesses in MountState for transitions. It also fetches the scoped context from these Components to create transitions: https://fburl.com/diffusion/l5rlt0k9
We want to get the scoped context based on the global key without fetching it from Component, so we're exposing getComponentKeysNeedingPreviousRenderData to get the key where we need the scoped context.

Reviewed By: adityasharat

Differential Revision: D23348743

fbshipit-source-id: 429cb1d7f8bd58f15940b66a69f57b522c655f78