Pass in global key to get scoped context
Summary:
Refactor Component.getScopedContext(LayoutStateContext c) => Component.getScopedContext(LayoutStateContext c, String globalKey)

When ComponentsConfiguration.useStatelessComponent is enabled, we'll use the LayoutStateContext and the global key that was passed in to get the scoped ComponentContext, instead of getting the global key from the Component field.

This essentially means that ComponentContext can be moved out of Component.

Reviewed By: adityasharat

Differential Revision: D23348751

fbshipit-source-id: cded91e8ce32b6878eba0e6908631ae11cc311ac