Blocklist for select component pre-allocation
Summary: This diff adds blocklist for pre-allocation of certain components based on QE-supplied list. We would like to experiment with blocking pre-allocations on the UI thread for components that are already being pre-allocated on the background thread. This is to prevent redundant operations on the UI thread in attempt to improve touch responsiveness.

Reviewed By: mihaelao

Differential Revision: D24075169

fbshipit-source-id: 6191c6bc22ce71ce4a7a682f4cc6557e95e7de60