Add API to RecyclerCollectionEventsController to manually set the viewport
Summary:
Add an API to RecyclerCollectionEventsController that allows us to supply viewport visibility updates that originate outside of the RecyclerView hosting this RecyclerCollectionComponent.

This is an uncommon use case. If the layout containing the RecyclerSpec does not contain a recycler that fills the screen, and its position changes due to external scrolling in the parent layout, this API can
be used to inform the section of these visibility changes. Most use cases will not need this, as the recycler will handle viewport updates during scrolling.

Reviewed By: astreet

Differential Revision: D24289375

fbshipit-source-id: 53cd527001970ca9637bda6647e46a2f2b41ca7f