Fix material text center issue
Summary:
This is to fix the issue of initial placeholder not centered correctly vertically with MaterialTextInput.
The BoxBackgroundMode by default is 0 and doesn't work as it places the pre-float placeholder below the vertical center which pre-maturely makes space for the floating label.
I fixed the rendering by making the BoxBackgroundMode 2 which is with bounding box and it doesn't have the centering issue. I made the the box stroke width 0 to hide the box which we don't need. I also added additional vertical padding to add more space between the floating label and the input text so it looks normal.

Reviewed By: andrewpmsmith

Differential Revision: D24125035

fbshipit-source-id: d48694e501f6ab2717905a56e18ed735b27d4c6a