Update extensions after unmountAll is called
Summary:
After unmountAll is called on a MountState with extensions, all extensions are removed from the MountDelegate.
If that's followed by a new mount call with the same extensions, they won't be re-registered on the mount delegate because shouldUpdate would return false.
That will cause issues with mount reference counting inconsistencies between the extension and the MountDelegate.

This diff removes the MountDelegate from the extension when it needs to be unregistered, and ensures that after unregistering the extensions they get re-registered on the next mount.

Reviewed By: adityasharat

Differential Revision: D24418863

fbshipit-source-id: 55944fb8c334f04f4b85e62e1ece9ae76407e8e6