Fix MaterialInputText on focus event issue
Summary:
`MaterialTextInput` creates a new `EditText` on each mount. This is because we have to wait until the `EditText` has been styled before we can add it to the `TextInputLayout`.

The OnFocusChangeListener was being lost each time a new EditText was added.

(See D24445983 for a repo of the issue)

Reviewed By: pasqualeanatriello

Differential Revision: D24448303

fbshipit-source-id: ec2cb5b6f9ed4afdd10f30a45af285861290c748