Add static method for constructing ComponentContext for nested trees
Summary: Adding this so we have a consistent place to link back to the parent context the nested tree was created from (see next diff in the stack). This also makes it a bit more clear how to construct a ComponentContext for a nested tree in a way that automatically passes through treeprops and logtag.

Reviewed By: mihaelao

Differential Revision: D24447137

fbshipit-source-id: 890412004f11f1e0f70e635968a29f8b350442e0