Pass ExtensionState to MountExtensions
Summary:
Instead of keeping a MountExtension's stateful data on the instance itself, we can have the MountState keep it.
That way, it's easier to manage creating extension instances since we can reuse the same instance across MountStates. The MountState will keep the extension state that's relevant for it.

This diff introduces an ExtensionState which right now only holds on to the MountDelegate that an extension is registered with. In the next diffs, the state that's kept on the extensions will be moved to the ExtensionState.

This makes it easy to keep the extension stateful but doesn't enforce that the MountExtension implementations only use static methods / are stateful. Would be interested to hear if you have any ideas on how we can achieve that

Reviewed By: colriot

Differential Revision: D24448660

fbshipit-source-id: 495a52c41a80f80e2ca5b4df3e3946559aea1307