IncrementalMountBinder depends on input
Summary:
Right now, the IncrementalMountBinder needs access to the extension's input which will be moved out the IncrementalMountExtension and only passed in to extension callbacks.
This diff keeps the input on the IncrementalMountBinder directly and updates it when the extension receives a new input.

There's also the option of adding a method on RenderUnit for `hostsRenderTrees()` which only LithoRenderUnit would override to true - but not sure if there's a reason we shouldn't do that

Reviewed By: adityasharat

Differential Revision: D24448774

fbshipit-source-id: 8d92cdd00db6a71833cbde6a04d6fb77b9f2930a