Clear last mounted tree id when setting a null component tree
Summary:
We use the last component tree id to check if we need to compare trees in order to run animations. This creates an issue when moving a component tree from one litho view to an other.
This case was causing to restart an any animation added to a row that is set as a sticky header (where we use different LithoViews).

Reviewed By: astreet

Differential Revision: D22477874

fbshipit-source-id: d82aa8b4c8eac00b7e8b40b56a55324fa89f66f9