Don't use RenderUnit binders in Litho MountState
Summary:
Litho MountState does not call into RenderUnit binders. All the functionality that is passed through the LithoRenderUnit binders is supposed to cover for Litho specific stuff that RenderCore's MountState doesn't support.

We should never call LithoRenderUnit's binders from Litho MountState because that means we'd end up doing the same operations twice. For example, Litho MountState sets view attributes when an item is mounted. The LithoRenderUnit binders also do that when called because RenderCore MountState doesn't do that. So if we use Litho MountState and also call the binders we can end up setting view attributes twice.

Litho MountState has a mount() implementation which is supposed to be similar to RenderCore and which still needs to communicate with the extensions when an item is mounted or updated to perform this extra work. We don't want to call the binders on the RenderUnit so we are calling on the extensions directly to do that.

This is more of a cleanup to make it easier to understand when we should call on the extensions to simulate the binder calls in Litho. (for RenderCore, all that work would be covered by the Extension binders).

Discussed with pasqualeanatriello that the attach/detach binders that the extensions create should be replaced with direct callbacks on the extensions, so this will be simplified further in future diffs. For now, it should unblock another test on Litho + extensions

Reviewed By: nicous

Differential Revision: D24648552

fbshipit-source-id: 3f50e52c84684ede2a7d51a25e902d63ffec0ddd