Remove incremental mount extension binder
Summary:
MountExtension binders need access to stateful info to perform operations. Since they get created when a RenderUnit is created, this means that if we're using the same RenderTree on different LithoViews, we can end up with binders that hold on to the wrong stateful info associated with the binders.

The correct thing to do is to pass the state when the binders are invoked to avoid this. However we don't want to change the Extension API for this particular use case, so instead we're introducing new callbacks on the extensions for this purpose.

The existing Extension binders will be replaced with these callbacks.

Reviewed By: adityasharat

Differential Revision: D24650851

fbshipit-source-id: 1c13ec53b3cbd094fd391e97a821e47558884535