Recursively force unmounted hosts to be mounted
Summary:
In cases like items with negative margins, they can be mounted before their host so we need to force the host to mount if it wasn't already mounted.
This was already done, but we were only mounting the direct host - the host's host could still be unmounted and cause a crash.

Reviewed By: nicous

Differential Revision: D24735706

fbshipit-source-id: 2d8b7d7160510bef97e2be97abc9781392c210a0