Compare willRender base contexts when using cached layout
Summary:
In `Component.willRender`, we check that the android context on the scoped ComponentContext of the component is the same as the context being passed in to willRender.
As we move layout-scoped info from the Component and on the LayoutState, in willRender we won't be able to fetch neither the scoped context nor the global key from the Component anymore - we'll have to pass either the scoped context or the global key to get layout-scoped info and that's not accessible outside of the framework.

Since the cached InternalNode is what could be reused under a different android context, we can move the assertion to where the cache is being fetched.

Reviewed By: astreet

Differential Revision: D23962701

fbshipit-source-id: eb5a4681ee2ce3fa16c4236f1dae2c680d9de455