Use same check to decide whether to wait for layout or not
Summary:
See attached exception for more details. We're hitting this because we're returning early from LayoutStateFuture.runAndGet when a bg thread's async layout is already being computed by something else, but when checking for sanity of null LayoutState in calculateLayout we're crashing because the output is not null.
Using the same check here as we do in runAndGet to decide whether to return early, and adding more info in the crash message

Differential Revision: D24922921

fbshipit-source-id: b780d624dc259d029a6f9a99816da762ccb404f2