Fixes bug where InterleavedDispatchDraw skips trailing items
Summary:
`getMountItemCount()` returns the number of non null Mount Items in the host, which can be less than the length of the array of Mount Items. The InterleavedDispatchDraw iterates over the array and because the count can be less than the actual length the trailing items of the host may get skipped.

This change set fixes this by using the length of the array instead.

Reviewed By: mihaelao

Differential Revision: D24953716

fbshipit-source-id: 411a74d74be0725891227e00aef11385522bbbca