Use available global key for stateless component
Summary:
Adds a utility method to get a component's global key based on whether the stateless flag is enabled or not.
If it's enabled, we use the global key which is already available from the layout calculation.
If it's not enabled, it defaults to the existing implementation which queries the component for the global key.

Replaces all usages of component.getGlobalKey with this utility method wherever we have a global key we can use. (all other usages will get the global key from the context in the next diff).

Reviewed By: adityasharat

Differential Revision: D24888399

fbshipit-source-id: 4159098112f16d3f6d445d956827f513cd4896bc