Change how we check if we have an unmount delegate
Summary:
We need this for future diffs where we want to remove the LayoutOutput in the unbind method for disappearing animations. The mLockedDisappearingMountitems now is a map from RenderUnit to AnimatableItems that we can then use to get properties that we were getting from the LayoutOutput.
Note: One thing to note here is that we can't use the render unit id as a key for the map because this map will hold disappearing animatable items and live through different layout passes, in this scenario we can have a new render unit matching an old one's id.

Reviewed By: adityasharat

Differential Revision: D24136232

fbshipit-source-id: ee13d325f777c470ee5ca8ae7b947d9db3104b57