Mark non-mountable items as released during mount
Summary:
In IncrementalMountExtension, when we check if an item is mountable in the current state and decide it's not but the extension had acquired a reference on it, we want to release the reference and unmount it if the ref count is zero.
If the check happens during mount, we won't know if the item can be released until all the extensions have been consulted, so we need to delay releasing the item until after mount.

The problem with the current approach is that we're saving these items to be released in afterMount, but don't immediately release the ref count - that means that isMountable will return true when MountState invokes it, so items which would need to unmounted will go through updating instead.

Reviewed By: adityasharat

Differential Revision: D25060277

fbshipit-source-id: 6d1aed827212259441e3efed44b6cc3b92c3bfd3