Corrects incremental mount extensions mounting condition
Summary:
Only mount an item if it intersects the visible bounds should be mounted.

The following condition would incorrectly mount an item if the **bottom of the local visible rect** is less than the **top of the item**.

```
while (state.mPreviousBottomsIndex > 0
          && localVisibleRect.top
              < byBottomBounds.get(state.mPreviousBottomsIndex - 1).getBounds().bottom) {
  ...
}
```

This change set fixes this issue by adding the additional check.

Reviewed By: mihaelao

Differential Revision: D25120965

fbshipit-source-id: 5e32937c2181ca1c32ea9ed9210c7f103a46273b