Change how we handle clip children for Animations.
Summary:
We need to set clip children to false when running animations. This was done on the Component Host and now we move this functionality to the transition manager. The key here is to keep track of what was the previos clip configuration for that view.
I changed the Host class to keep track of the clip configuration for the view as the getClipChildren method is only api level 18 +.

Reviewed By: adityasharat

Differential Revision: D25128914

fbshipit-source-id: 4b0e9fab16a16157747e16b941341550c0dc9c47