Unmounted host crash in RenderCore
Summary:
Found another situation where we could be mounting an item when its host hasn't mounted.
Consider the following mount hierarchy:

RootHost: parent null, Rect (0, 0, 10, 20)
Host1: parent RootHost, Rect (0, 10, 10, 20)
MountableItem1: parent Host1 , Rect(0, 8, 10, 20)
Host2: parent RootHost

1) Mount with Rect (0, 0, 10, 9)
Host1 is not visible but gets force-mounted because MountableItem1 is in the visible rect.

2) Another mount moves Host1 under a different host, say Host2, then Host1 and its children get unmounted in prepareMount
When checking if items are mountable again, Host1 is still out of bounds so not mountable, and MountableItem1 is mountable. However the reference was already acquired for the item, so in maybeAcquireReference we don't ensure the host is also mounted.

The test covers this scenario, it was failing before the change and is passing now

Reviewed By: adityasharat

Differential Revision: D25273283

fbshipit-source-id: 6a36c1dc12051c7fd0d759f6886eb3d01a9a15d7